update dtype in docstring

+++ src/ansys/dpf/core/elements.py
@@ -669,7 +669,7 @@ class Elements:
         >>> field = vol.outputs.fields_container()[0]
         >>> ind, mask = elements.map_scoping(field.scoping)
         >>> ind
-        array([0, 1, 2, 3, 4, 5, 6, 7])
+        array([0, 1, 2, 3, 4, 5, 6, 7], dtype=int32)
 
         """
         if external_scope.location in ["Nodal", "NodalElemental"]:
